+++ CHANGELOG.md
@@ -11,6 +11,7 @@
       It replaces `Modifier::Attributes`.
     * `math::Aab::expand_or_shrink()`
     * `math::Vol::origin_empty()`, infallible construction of empty volume.
+    * `physics::Contact::aab()`
     * `time::Clock::phase()`, returns the clock’s phase, which was previously only obtainable through `Tick`.
     * `time::Tick::{delta_t_f64, delta_t_ps64}()`, convenience for common conversions.